pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vitia-invenire"
version = "0.1.0"
description = "Windows Supply Chain Security Assessment Tool"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "click>=8.0",
    "rich>=13.0",
    "jinja2>=3.0",
    "pyyaml>=6.0",
    "psutil>=5.9",
    "pefile>=2023.2.7",
    "signify>=0.7.0",
]

[project.optional-dependencies]
nmap = ["python-nmap>=0.7"]
nessus = ["pyTenable>=1.4"]
virustotal = ["requests>=2.28"]
cpuinfo = ["py-cpuinfo>=9.0"]
all = [
    "python-nmap>=0.7",
    "pyTenable>=1.4",
    "requests>=2.28",
    "py-cpuinfo>=9.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
vitia-invenire = "vitia_invenire.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
